+++ src/main/java/blackjack/domain/player/BettingAmounts.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 import java.util.List;
