+++ src/main/java/blackjack/domain/result/Result.java
@@ -1,6 +1,6 @@
 package blackjack.domain.result;
 
-import blackjack.domain.Profit;
+import blackjack.domain.player.Profit;
 import blackjack.domain.player.Name;
 import java.util.Map;
