+++ src/test/java/blackjack/domain/BlackjackTest.java
@@ -5,11 +5,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import blackjack.domain.card.CardValue;
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Deck;
+import blackjack.domain.player.BettingAmounts;
 import blackjack.domain.player.Dealer;
 import blackjack.domain.player.GamePlayer;
 import blackjack.domain.player.Name;
 import blackjack.domain.player.Names;
 import blackjack.domain.player.Player;
+import blackjack.domain.player.Profit;
 import blackjack.fixture.CardFixture;
 import blackjack.fixture.PlayerFixture;
 import java.util.List;
